offline/requirements.txt
marker-pdf>=1.10.0
ollama>=0.4.0
rich>=13.0
PyMuPDF>=1.24.0
